delete old streaming compression interfaces
