--- compression/src/main/java/net/daporkchop/lib/compression/context/DCtx.java
@@ -1,57 +1,0 @@
-/*
- * Adapted from The MIT License (MIT)
- *
- * Copyright (c) 2018-2020 DaPorkchop_
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation
- * files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy,
- * modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software
- * is furnished to do so, subject to the following conditions:
- *
- * Any persons and/or organizations using this software must include the above copyright notice and this permission notice,
- * provide sufficient credit to the original authors of the project (IE: DaPorkchop_), as well as provide a link to the original project.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
- * OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS
- * BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- *
- */
-
-package net.daporkchop.lib.compression.context;
-
-import io.netty.buffer.ByteBuf;
-import lombok.NonNull;
-import net.daporkchop.lib.compression.util.exception.DictionaryNotAllowedException;
-import net.daporkchop.lib.natives.util.exception.InvalidBufferTypeException;
-
-/**
- * A context for doing repeated one-shot compression operations.
- *
- * @author DaPorkchop_
- */
-public interface DCtx extends Context {
-    /**
-     * Convenience method, equivalent to {@code decompress(src, dst, null);}.
-     *
-     * @see #decompress(ByteBuf, ByteBuf, ByteBuf)
-     */
-    default boolean decompress(@NonNull ByteBuf src, @NonNull ByteBuf dst) throws InvalidBufferTypeException {
-        return this.decompress(src, dst, null);
-    }
-
-    /**
-     * Decompresses the given compressed data into the given destination buffer.
-     * <p>
-     * If the destination buffer does not have enough space writable for the decompressed data, the operation will fail and both buffer's indices will remain
-     * unchanged, however the destination buffer's contents may be modified.
-     * <p>
-     * In either case, the indices of the dictionary buffer remain unaffected.
-     *
-     * @param src the {@link ByteBuf} to read compressed data from
-     * @param dst the {@link ByteBuf} to write decompressed data to
-     * @param dict the (possibly {@code null}) {@link ByteBuf} containing the dictionary to be used for decompression
-     * @return whether or not decompression was successful. If {@code false}, the destination buffer was too small for the decompressed data
-     */
-    boolean decompress(@NonNull ByteBuf src, @NonNull ByteBuf dst, ByteBuf dict) throws InvalidBufferTypeException, DictionaryNotAllowedException;
-}
