+++ compression/src/main/java/net/daporkchop/lib/compression/context/PDeflater.java
@@ -22,55 +22,42 @@ package net.daporkchop.lib.compression.context;
 
 import io.netty.buffer.ByteBuf;
 import lombok.NonNull;
-import net.daporkchop.lib.compression.util.exception.ContextFinishedException;
-import net.daporkchop.lib.compression.util.exception.ContextFinishingException;
+import net.daporkchop.lib.compression.util.exception.DictionaryNotAllowedException;
 import net.daporkchop.lib.natives.util.exception.InvalidBufferTypeException;
 
 /**
- * Deflates (compresses) data.
+ * A context for doing repeated one-shot compression operations.
  *
  * @author DaPorkchop_
  */
-public interface PDeflater extends StreamingContext<PDeflater> {
+public interface PDeflater extends Context {
     /**
-     * Deflates the given source data into the given destination buffer.
-     * <p>
-     * Rather than expanding the destination buffer if needed, this method will simply abort compression if not enough space is available. In such a case the
-     * reader/writer indices of both buffers will remain unaffected, however the contents of the destination buffer may be modified.
-     * <p>
-     * This method will implicitly reset the context before the actual compression. Any previous state will be ignored.
+     * Convenience method, equivalent to {@code compress(src, dst, null);}.
      *
-     * @param src the {@link ByteBuf} to read data from
-     * @param dst the {@link ByteBuf} to write data to
-     * @return whether or not there was enough space in the destination buffer for the compressed data
+     * @see #compress(ByteBuf, ByteBuf, ByteBuf)
      */
-    boolean fullDeflate(@NonNull ByteBuf src, @NonNull ByteBuf dst) throws InvalidBufferTypeException;
-
-    /**
-     * Deflates the given source data into the given destination buffer.
-     * <p>
-     * This will continually grow the destination buffer until there is enough space for deflation to be finished successfully.
-     * <p>
-     * This method will implicitly reset the context before the actual compression. Any previous state will be ignored.
-     *
-     * @param src the {@link ByteBuf} to read data from
-     * @param dst the {@link ByteBuf} to write data to
-     */
-    default void fullDeflateGrowing(@NonNull ByteBuf src, @NonNull ByteBuf dst) throws InvalidBufferTypeException {
-        this.reset().src(src).dst(dst);
-
-        do {
-            this.update(false);
-        } while (src.isReadable() && dst.ensureWritable(8192).isWritable());
-
-        while (!this.finish()) {
-            dst.ensureWritable(8192);
-        }
+    default boolean compress(@NonNull ByteBuf src, @NonNull ByteBuf dst) throws InvalidBufferTypeException {
+        return this.compress(src, dst, null);
     }
 
-    @Override
-    PDeflater update(boolean flush) throws ContextFinishedException, ContextFinishingException;
+    /**
+     * Compresses the given source data into the given destination buffer at the configured compression level.
+     * <p>
+     * If the destination buffer does not have enough space writable for the compressed data, the operation will fail and both buffer's indices will remain
+     * unchanged, however the destination buffer's contents may be modified.
+     * <p>
+     * In either case, the indices of the dictionary buffer remain unaffected.
+     *
+     * @param src  the {@link ByteBuf} to read source data from
+     * @param dst  the {@link ByteBuf} to write compressed data to
+     * @param dict the (possibly {@code null}) {@link ByteBuf} containing the dictionary to be used for compression
+     * @return whether or not compression was successful. If {@code false}, the destination buffer was too small for the compressed data
+     * @throws DictionaryNotAllowedException if the dictionary buffer is not {@code null} and this context does not allow use of a dictionary
+     */
+    boolean compress(@NonNull ByteBuf src, @NonNull ByteBuf dst, ByteBuf dict) throws InvalidBufferTypeException, DictionaryNotAllowedException;
 
-    @Override
-    boolean finish() throws ContextFinishedException;
+    /**
+     * @return the configured compression level
+     */
+    int level();
 }
