--- compression/src/main/java/net/daporkchop/lib/compression/provider/OneShotCompressionProvider.java
@@ -1,55 +1,0 @@
-/*
- * Adapted from The MIT License (MIT)
- *
- * Copyright (c) 2018-2020 DaPorkchop_
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation
- * files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy,
- * modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software
- * is furnished to do so, subject to the following conditions:
- *
- * Any persons and/or organizations using this software must include the above copyright notice and this permission notice,
- * provide sufficient credit to the original authors of the project (IE: DaPorkchop_), as well as provide a link to the original project.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
- * OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS
- * BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- *
- */
-
-package net.daporkchop.lib.compression.provider;
-
-import net.daporkchop.lib.compression.context.CCtx;
-import net.daporkchop.lib.compression.context.DCtx;
-import net.daporkchop.lib.compression.util.exception.InvalidCompressionLevelException;
-
-/**
- * An implementation of a compression algorithm that supports one-shot compression via {@link CCtx} and {@link DCtx}.
- *
- * @author DaPorkchop_
- */
-public interface OneShotCompressionProvider extends CompressionProvider {
-    /**
-     * Creates a new {@link CCtx} with the default compression level.
-     *
-     * @see #compressionContext(int)
-     */
-    default CCtx compressionContext() {
-        return this.compressionContext(this.levelDefault());
-    }
-
-    /**
-     * Creates a new {@link CCtx} with the given compression level.
-     *
-     * @param level the compression level to use
-     * @return a new {@link CCtx} with the given compression level
-     * @throws InvalidCompressionLevelException if the given compression level is invalid
-     */
-    CCtx compressionContext(int level) throws InvalidCompressionLevelException;
-
-    /**
-     * @return a new {@link DCtx}
-     */
-    DCtx decompressionContext();
-}
