--- compression/src/main/java/net/daporkchop/lib/compression/util/StreamingWrapperCCtx.java
@@ -1,104 +1,0 @@
-/*
- * Adapted from The MIT License (MIT)
- *
- * Copyright (c) 2018-2020 DaPorkchop_
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation
- * files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy,
- * modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software
- * is furnished to do so, subject to the following conditions:
- *
- * Any persons and/or organizations using this software must include the above copyright notice and this permission notice,
- * provide sufficient credit to the original authors of the project (IE: DaPorkchop_), as well as provide a link to the original project.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
- * OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS
- * BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- *
- */
-
-package net.daporkchop.lib.compression.util;
-
-import io.netty.buffer.ByteBuf;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
-import lombok.experimental.Accessors;
-import net.daporkchop.lib.compression.context.CCtx;
-import net.daporkchop.lib.compression.context.PDeflater;
-import net.daporkchop.lib.compression.provider.StreamingCompressionProvider;
-import net.daporkchop.lib.compression.util.exception.DictionaryNotAllowedException;
-import net.daporkchop.lib.compression.util.exception.InvalidCompressionLevelException;
-import net.daporkchop.lib.natives.util.exception.InvalidBufferTypeException;
-import net.daporkchop.lib.unsafe.util.AbstractReleasable;
-
-/**
- * An implementation of {@link CCtx} that wraps a {@link PDeflater}.
- *
- * @author DaPorkchop_
- */
-@RequiredArgsConstructor(access = AccessLevel.PROTECTED)
-@Accessors(fluent = true)
-public class StreamingWrapperCCtx extends AbstractReleasable implements CCtx {
-    @Getter
-    @NonNull
-    protected final StreamingCompressionProvider provider;
-    @NonNull
-    protected final PDeflater           deflater;
-    @Getter
-    protected final int                 level;
-
-    public StreamingWrapperCCtx(@NonNull StreamingCompressionProvider provider, int level) throws InvalidCompressionLevelException {
-        this(provider, provider.deflater(level), level);
-    }
-
-    @Override
-    public boolean compress(@NonNull ByteBuf src, @NonNull ByteBuf dst) throws InvalidBufferTypeException {
-        return this.deflater.reset().fullDeflate(src, dst);
-    }
-
-    @Override
-    public boolean compress(@NonNull ByteBuf src, @NonNull ByteBuf dst, ByteBuf dict) throws InvalidBufferTypeException, DictionaryNotAllowedException {
-        if (dict != null && !this.hasDict()) {
-            throw new DictionaryNotAllowedException();
-        }
-
-        this.deflater.reset();
-        if (dict != null) {
-            this.deflater.dict(dict);
-        }
-        return this.deflater.fullDeflate(src, dst);
-    }
-
-    @Override
-    public boolean hasDict() {
-        return this.deflater.hasDict();
-    }
-
-    @Override
-    public boolean directAccepted() {
-        return this.deflater.directAccepted();
-    }
-
-    @Override
-    public boolean heapAccepted() {
-        return this.deflater.heapAccepted();
-    }
-
-    @Override
-    public boolean isAcceptable(@NonNull ByteBuf buf) {
-        return this.deflater.isAcceptable(buf);
-    }
-
-    @Override
-    public ByteBuf assertAcceptable(@NonNull ByteBuf buf) throws InvalidBufferTypeException {
-        return this.deflater.assertAcceptable(buf);
-    }
-
-    @Override
-    protected void doRelease() {
-        this.deflater.release();
-    }
-}
